Change project architecture
