--- shan_frame/code_generator.py
@@ -1,5 +1,0 @@
-from .ir.model import Model
-
-
-def generate_code(output_dir: str, model: Model) -> None:
-    raise NotImplementedError("code_generator.generate_code")
