+++ shan_frame/code_generator/__init__.py
@@ -1,0 +1,9 @@
+from ir import IRModel
+
+
+class CodeGenerator:
+    def __init__(self, output_dir: str) -> None:
+        pass
+
+    def generate(self, model: IRModel) -> None:
+        raise NotImplementedError("CodeGenerator.generate()")
