--- shan_frame/ir/layer_parser/__init__.py
@@ -1,1 +1,0 @@
-from .conv_2d import parse_conv_2d
