--- shan_frame/target_arch.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class TargetArch(Enum):
-    CORTEX_M7 = 0
-    CORTEX_M4 = 1
